Android accelerometer function
Uses the Accelerometer and Orientation sensor to determine the maximum
acceleration from one movement and then outputs this to logcat. Menu
button allows the calibration of the accelerometer for all axis as well
as the resetting of the calibration and the movement counter.

Basic UI which shows the current readings for the accelerometer in X,
Y, and Z, total acceleration, current orientation, and number of
movements. A start/stop button controls the output to logcat.

Utilises a wakelock function to prevent the screen turning off when
using the app and will disply a warning if either the accelerometer or
orientation sensors are not available.