Added playing of pre-loaded pcm to trunk
This works using a thread called tToneLoad to load a 16-bit signed
PCM/WAV file and add it to a linked-list of samples, which a second
thread, tToneBuffer adds to the hardware audio buffer via the
AudioOutputManager class. Playback of this buffer is controlled using a
Start/Pause button and a Stop button.

Code also exists in this release to generate a tone using a third
thread, tToneGenerate. This is currently not hooked up to any button
listeners as there are issues with playback.