Updated to use new SoundManager
This sound smooth for the first few seconds of playback and then goes a
bit stuttery. There is a slight delay between pressing Start music and
the audio beginning due to the time taken to both fill the audio buffer
beyond its minimum size, and to realise that this is the case and play
the audio.