Changed FilterManager to add Oscillators
FilterManager can now have oscillators attached to filters and will use
them when this is the case. Sample rate is now handled by the Filter
superclass, as is the abstract method that applies a filter using an
oscillator - which will replace the internally stored cutoff/variables.
The depth LFO is now using the AmplitudeFilter class.